BACK_END/requirements.txt
flask
flask_restful
flask_session
flask_cors
requests
opencv-python
opencv-contrib-python
numpy
azure-storage-blob
msrest
azure-cognitiveservices-vision-computervision
azure-cognitiveservices-vision-face
